Add `Calc.get()` and other Calc changes
- Introducing `Calc.get()`
`Calc.get()` is a core function which checks the attacker and defender's
ability and items, as well as the move and field conditions, for
anything that may alter the portion of the damage formula passed into
this method.

- Add `Calc.calcing` property. This is true if we are running a
calculation. See: `Calc.calcDamageNumbers()`
- Rename `Calc.getDamageNumbers()` to `Calc.calcDamageNumbers()`.
`Calc.get*()` is now reserved for functions centered around `Calc.get()`